# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Stellar Trust Escrow API
  version: 0.1.0
  description: |
    REST API for the Stellar Trust Escrow platform — a trust-minimised escrow
    service built on the Stellar blockchain.

    ## Authentication
    Most endpoints require a Bearer JWT token in the `Authorization` header.
    Admin endpoints additionally require the `x-admin-api-key` header.

    ## Pagination
    All collection endpoints use **cursor-based pagination** and return:
    ```
    {
      "data": [...],
      "next_cursor": "<opaque-base64url-token>",
      "has_more": true
    }
    ```
    Pass `?cursor=<next_cursor>` to fetch the next page. The cursor encodes the
    sort key and direction and is stable across concurrent inserts.

    ## Idempotency Keys
    All `POST` and `PATCH` endpoints accept an optional `Idempotency-Key` header.
    Supplying this header makes the call safe to retry:
    - The response is cached in Redis for **24 hours** under `(tenantId, method, path, key)`.
    - Duplicate requests within the TTL return the cached response immediately
      (status `X-Idempotency-Replayed: true`).
    - If a duplicate arrives while the first is still in-flight, a **409 Conflict** is returned.
    - Use any unique string (UUID recommended): `Idempotency-Key: <uuid-v4>`

    ## Environments
    - **Testnet**: `https://api.test.stellar-trust-escrow.example.com`
    - **Mainnet**: `https://api.stellar-trust-escrow.example.com`
  contact:
...
  license:
    name: AGPL-3.0
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token using refresh token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: New access token returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: integer
                    example: 3600
        '401':
          description: Refresh token expired or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/users/{address}:
    get:
      tags: [Users]
      summary: Get user profile (reputation, escrow history, stats)
      operationId: getUserProfile
      security:
        - BearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          description: Stellar public key (G...)
          schema:
            type: string
            example: GDKHX...QZ4
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/users/{address}/escrows:
    get:
      tags: [Users]
      summary: List escrows for a specific user
      operationId: getUserEscrows
      security:
        - BearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: role
          in: query
          schema:
            type: string
            enum: [client, freelancer, all]
            default: all
        - name: status
          in: query
          schema:
            type: string
            enum: [Active, Completed, Disputed, Cancelled]
      responses:
        '200':
          description: Paginated user escrow list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEscrows'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/users/{address}/stats:
    get:
      tags: [Users]
      summary: Aggregated user statistics
      operationId: getUserStats
      security:
        - BearerAuth: []
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/auth/login
    AdminAuth:
      type: apiKey
      in: header
      name: x-admin-api-key
      description: Admin API key for protected admin endpoints
  parameters:
    Page:
      name: page
      in: query
      description: 1-based page number. Values below 1 are normalised to 1.
      schema:
        type: integer
        default: 1
        minimum: 1
    Limit:
      name: limit
      in: query
      description: Page size. Defaults to 20 and is capped at 100.
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Something went wrong
        errorId:
          type: string
          description: Sentry event ID for support correlation
    PaginatedResponse:
      type: object
      required: [data, page, limit, total, totalPages, hasNextPage, hasPreviousPage]
      properties:
        data:
          type: array
          items: {}
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3
        hasNextPage:
          type: boolean
          example: true
        hasPreviousPage:
          type: boolean
          example: false
...
    Escrow:
      type: object
      properties:
        id:
          type: string
          example: escrow_abc123
        client:
          type: string
          example: GDKHX...QZ4
        freelancer:
          type: string
          example: GDKHX...QZ4
        totalAmount:
          type: string
          example: '1000'
        status:
          type: string
          enum: [Active, Completed, Disputed, Cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginatedEscrows:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Escrow'
...
    UserProfile:
      type: object
      properties:
        address:
          type: string
        email:
          type: string
          nullable: true
        kycStatus:
          type: string
          enum: [Pending, Init, Processing, Approved, Declined]
        reputation:
          type: object
          properties:
            score:
              type: number
            totalReviews:
              type: integer
            completionRate:
              type: number
        escrowStats:
          type: object
          properties:
            asClient:
              type: integer
            asFreelancer:
              type: integer
            completed:
              type: integer
            disputed:
              type: integer
        createdAt:
          type: string
          format: date-time
    UserStats:
      type: object
      properties:
        address:
          type: string
        totalVolume:
          type: string
          description: Total XLM volume
        completionRate:
          type: number
          example: 0.92
        avgMilestoneTimeDays:
          type: number
        totalEscrows:
          type: integer
        completedEscrows:
          type: integer
        disputedEscrows:
          type: integer
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...